producer/requirements.txt
Flask
azure-servicebus